feat(core): Move `EqHash` to core and rename
