+++ core/src/mast/mod.rs
@@ -7,7 +7,11 @@ use core::{
     ops::{Index, IndexMut},
 };
 
-use miden_crypto::hash::rpo::RpoDigest;
+use miden_crypto::hash::{
+    blake::{Blake3Digest, Blake3_256},
+    rpo::RpoDigest,
+    Digest,
+};
 
 mod node;
 pub use node::{
@@ -563,6 +567,170 @@ impl fmt::Display for DecoratorId {
 impl Serializable for DecoratorId {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.0.write_into(target)
+    }
+}
+
+// MAST NODE EQUALITY
+// ================================================================================================
+
+// TODO: We need to export this so we can use it in MastForestBuilder, but do we really want this to
+// be public? Alternatively we can move move MastForestBuilder into core as well, but this might not
+// be a good fit since it is only used during assembly.
+
+/// Represents the hash used to test for equality between [`MastNode`]s.
+///
+/// The decorator root will be `None` if and only if there are no decorators attached to the node,
+/// and all children have no decorator roots (meaning that there are no decorators in all the
+/// descendants).
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+pub struct MastNodeEq {
+    mast_root: RpoDigest,
+    decorator_root: Option<Blake3Digest<32>>,
+}
+
+// TODO: Document public functions and assumptions about forest and the index map.
+impl MastNodeEq {
+    pub fn new(mast_root: RpoDigest) -> Self {
+        Self { mast_root, decorator_root: None }
+    }
+
+    pub fn with_decorator_root(mast_root: RpoDigest, decorator_root: Blake3Digest<32>) -> Self {
+        Self {
+            mast_root,
+            decorator_root: Some(decorator_root),
+        }
+    }
+
+    pub fn from_mast_node(
+        forest: &MastForest,
+        hash_by_node_id: &BTreeMap<MastNodeId, MastNodeEq>,
+        node: &MastNode,
+    ) -> MastNodeEq {
+        match node {
+            MastNode::Block(node) => {
+                let mut bytes_to_hash = Vec::new();
+
+                for &(idx, decorator_id) in node.decorators() {
+                    bytes_to_hash.extend(idx.to_le_bytes());
+                    bytes_to_hash.extend(forest[decorator_id].eq_hash().as_bytes());
+                }
+
+                // Add any `Assert` or `U32assert2` opcodes present, since these are not included in
+                // the MAST root.
+                for (op_idx, op) in node.operations().enumerate() {
+                    if let Operation::U32assert2(inner_value)
+                    | Operation::Assert(inner_value)
+                    | Operation::MpVerify(inner_value) = op
+                    {
+                        let op_idx: u32 = op_idx
+                            .try_into()
+                            .expect("there are more than 2^{32}-1 operations in basic block");
+
+                        // we include the opcode to differentiate between `Assert` and `U32assert2`
+                        bytes_to_hash.push(op.op_code());
+                        // we include the operation index to distinguish between basic blocks that
+                        // would have the same assert instructions, but in a different order
+                        bytes_to_hash.extend(op_idx.to_le_bytes());
+                        bytes_to_hash.extend(inner_value.to_le_bytes());
+                    }
+                }
+
+                if bytes_to_hash.is_empty() {
+                    MastNodeEq::new(node.digest())
+                } else {
+                    let decorator_root = Blake3_256::hash(&bytes_to_hash);
+                    MastNodeEq::with_decorator_root(node.digest(), decorator_root)
+                }
+            },
+            MastNode::Join(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[node.first(), node.second()],
+                node.digest(),
+            ),
+            MastNode::Split(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[node.on_true(), node.on_false()],
+                node.digest(),
+            ),
+            MastNode::Loop(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[node.body()],
+                node.digest(),
+            ),
+            MastNode::Call(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[node.callee()],
+                node.digest(),
+            ),
+            MastNode::Dyn(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[],
+                node.digest(),
+            ),
+            MastNode::External(node) => eq_hash_from_parts(
+                forest,
+                hash_by_node_id,
+                node.before_enter(),
+                node.after_exit(),
+                &[],
+                node.digest(),
+            ),
+        }
+    }
+}
+
+fn eq_hash_from_parts(
+    forest: &MastForest,
+    hash_by_node_id: &BTreeMap<MastNodeId, MastNodeEq>,
+    before_enter_ids: &[DecoratorId],
+    after_exit_ids: &[DecoratorId],
+    children_ids: &[MastNodeId],
+    node_digest: RpoDigest,
+) -> MastNodeEq {
+    let pre_decorator_hash_bytes =
+        before_enter_ids.iter().flat_map(|&id| forest[id].eq_hash().as_bytes());
+    let post_decorator_hash_bytes =
+        after_exit_ids.iter().flat_map(|&id| forest[id].eq_hash().as_bytes());
+
+    // Reminder: the `EqHash`'s decorator root will be `None` if and only if there are no
+    // decorators attached to the node, and all children have no decorator roots (meaning that
+    // there are no decorators in all the descendants).
+    if pre_decorator_hash_bytes.clone().next().is_none()
+        && post_decorator_hash_bytes.clone().next().is_none()
+        && children_ids
+            .iter()
+            .filter_map(|child_id| hash_by_node_id[child_id].decorator_root)
+            .next()
+            .is_none()
+    {
+        MastNodeEq::new(node_digest)
+    } else {
+        let children_decorator_roots = children_ids
+            .iter()
+            .filter_map(|child_id| hash_by_node_id[child_id].decorator_root)
+            .flat_map(|decorator_root| decorator_root.as_bytes());
+        let decorator_bytes_to_hash: Vec<u8> = pre_decorator_hash_bytes
+            .chain(post_decorator_hash_bytes)
+            .chain(children_decorator_roots)
+            .collect();
+
+        let decorator_root = Blake3_256::hash(&decorator_bytes_to_hash);
+        MastNodeEq::with_decorator_root(node_digest, decorator_root)
     }
 }
